test(auditor): move shared helpers from suite conftest to fixtures/auditor.py
Mirrors the existing fixtures/audit.py / fixtures/auth.py / fixtures/signoz.py
layout — domain-named module under fixtures/ registered via pytest_plugins,
suite-local conftest.py kept thin (only the signoz fixture override that
configures the auditor container).
